More whitespace fixes
* Paul E. McKenney <paulmck@linux.vnet.ibm.com> wrote:

> The other thing that I am working on to (hopefully) help out is
> at:
>
> git://git.kernel.org/pub/scm/linux/kernel/git/paulmck/perfbook.git
>
> This is still an early draft, and there are still some gaping
> holes, but making progress, and have gotten some good feedback
> from a few people who have gone through it.  Even got one
> contribution -- just a typo fix, but one has to start somewhere.

Wow, that's really impressive!

 August 5, 2006: 76 pages
 ...
 February 28, 2009: 292 pages

Let me be the second contributor of a typo fix. (I also fixed all
occurances of trailing whitespaces, via a script. My editor shows
them in red.) Find the diff further below.

It's way too big for me to read fully right now, but the intro pages
i have read i enjoyed a lot :)

This book would IMHO be a worthy addition to the upstream kernel in
terms of Documentation/perfbook.txt - or perhaps in terms of
Documentation/perfbook/ with the .tex files and small diagrams. The
full material is too big, unfortunately. But it would be nice to
make sure aspiring parallel programmers would be more aware of it.

> So, why waste time on a book rather than just directly
> contributing? Especially given that it is a freely downloadable
> book, which guarantees that no publisher will ever touch it?
>
> Well, even though I am a bit younger than Hugh Dickins and Jack
> Steiner, I did turn 50 last year.  Judging from my grandparents, I
> have somewhere between five and thirty years left during which I
> will be able to contribute at something vaguely resembling my
> current level.  So, over the past decade, my priority has shifted
> from contributing solely with my own eyes and hands to increasing
> the abilities of others -- while still getting some hacking in
> myself.
>
> Actually, beyond a certain point, hacking is just as necessary to
> my well-being as is sleeping, so I guess that it is not like I
> have a choice.  ;-)
>
> All that aside, given your encouragement, I will look into seeing
> if I can do better filtering in gmail (excluding messages in which
> "RCU" appears only as part of CONFIG_CLASSIC_RCU=y, for example)
> and if I can free up more time to reveiw and hack.  Some weeks
> will be better than others in that regard.
>
> 							Thanx, Paul

I'm sure it will be welcome. AFAICS people on lkml generally reacted
positively to any analysis you posted in various threads - even if
RCU was just a tangential entry point into the discussion.

Leaning back and thinking things through is something that is pretty
hard to do to developers of fresh new code - bugs are fresh in mind
and unimplemented features are desired. So we have a constant (and
growing!) problem with review depth on lkml.

	Ingo

Signed-off-by: Ingo Molnar <mingo@elte.hu>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>