Typo fixes for the cpu directory
Hi,

Here are fixes for the typos that I found in Chapter 2.  I'm not sure
that all of them are correct.

--
Matt                                           https://ftbfs.org/kraai

>From b7f90d0625c81654b0b38292512be916840cc6d7 Mon Sep 17 00:00:00 2001
From: Matt Kraai <kraai@ftbfs.org>
Date: Mon, 10 Jan 2011 05:42:11 -0800
Subject: [PATCH] Fix typos in the cpu directory.

Change "exponentially" to "exponential", remove "are now readily
available", change "go" to "goes", "nanosecond" to "nanoseconds", and
"lock" to "clock", add "the", change "Shared memory" to
"Shared-memory", add "systems", and change a comma to a period.

Signed-off-by: Matt Kraai <kraai@ftbfs.org>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>