Remove extra brace
Hi,
as you can guess, I'm not a kernel developer and I'm not confident with the
way patches are submitted.
I've tried to follow the FAQ.txt file and to read some documents, but I have
not found where I shoud submit my patches, so I'm sending it to your address.
Here's my _minor_ contribution to the excellent book

regards,
Davide

Signed-off-by: Davide Viti <zinosat@tiscali.it>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>