Make HTMovf use adaptive approach, computing most efficient way
If the cache is nearly empty, it is more efficient to count the number
of overflow (failure) configurations.  If the cache is nearly full,
it is more efficient to count the number of non-overflow (success)
configurations.  So first count the applicable number of computations,
then use the fastest one.

This works because most of the overhead is in maxima's indefinite-precision
computations.

Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>