qqz: Cross-link questions and answers
This is useful even if you _are_ actually solving the quick quizzes.
Some ebook readers in particular support hyperlinks but have very
limited support for PDF search and back/forward.

The implementation is easy; it works by adding QQ.chapter.quiz labels
at the question and QQA.chapter.quiz at the answer.  The only
interesting point is using \edef and \getrefnumber to extract the
chapter number for answers, in a way that can be passed to
\hyperref.  An alternative would be to just use consecutive numbering
for the QQ and QQA labels, but I find this a little more elegant.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>