cpu: Rearrange summary of chapter
Summary of chapter and introduction about following chapters is in the
end part of `swdesign.tex` though it is not software design specific
content.  Moreover, the detailed introduction about following chapters
is in `cpu.tex`, too.  It would be better to gather them in one `tex`
file.  This commit, which is pure code motion, gathers them in `cpu.tex`.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>