whymb: Remove ARM's short multi processor history description
A description about ARM processor says that history of multi processor
ARM CPU is about five years.  However, the sentence was written in
2010 by commit 864762cb5206f31b71757e4da8362d8c1c0e3b7c ("Add ARM to the
"why memory barriers" section.").  It's 2016 now.  Multi processor ARM
CPUs are common and its history is more than a decade.  The sentence can
be simply modified to say ten years rather than five years.  However, it
would be better to simply remove the sentence because  modifying the
word every year would be painful and a decade in computer industry is
not a short period though it may be arguable.  For the reason, this
commit removes the sentence.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>