advsync: Add necessary blank lines around minipages
In the `Pair-Wise Memory Barriers: Portable Combinations`
subsubsection, we need blank lines before and after minipages
that embody code snippet tables. Without them, minipages are
forced to be placed in strange ways. This commit adds those
necessary blank lines.
Also it adds "\scriptsize" commands to the minipages for their
consistency with the one in the previous subsubsection.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>