qqz: Improve accuracy of cross-links
In commit 33b93f8258f5 ("qqz: Cross-link questions and answers"),
cross-links of quick quizzes and their answers were implemented.
However it uses the \stepcounter{} command, and direct references
of 'quickquizctr' counter is not output to the .aux file.
The symptom is when you click on a hyperlink in the resulting PDF,
you jump to a random nearby label such as that of Figure, Table,
or Section, etc. which happens just before the \label{} using
'quickquizctr'.

By using \refstepcounter{} instead and using 2 counters
'quickquizctr' and 'quickquizctrC' ('C' stands for 'Chapter'),
each of which is automatically reset at a new chapter and
a new section respectively, and defining \theHNum at the head
of qqz.sty, then redefining it within the \QuickQuizAnsers
command, matching QQ.n.m's and QQA.n.m's are output to the .aux
file correctly.

By applying this change, resulting hyperlinks become to jump to
places almost exactly where you'd expect.

This change may have the effect of increasing the required number
of 'pdflatex' iterations to get a final result. The following
commit will take care of it.

This commit also updates the copyright notices in the header,
including the name of Paolo Bonzini who authored the commit
mentioned above.

Signed-off-by: Akira Yokosawa <akiysw@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>