Another regression fix of build script improvement
After commit 4038e781ed56 ("Improve behavior of build scripts"),
it is impossible to build 'perfbook-1c.pdf' and 'perfbook-hb.pdf'
just after cloning the git repository.
The reason is the lack of rules for 'perfbook-1c.tex' and
'perfbook-hb.tex'targets, which are required for them.
This commit adds those rules to Makefile.
The reason I couldn't find the bug during my test is both
'perfbook-1c.tex' and 'perfbook-hb.tex' survive even
'make neatfreak' which I assumed to be used to clean everything.
To avoid similar regression to recur, this commit also appends
them to the 'clean' target.

Signed-off-by: Akira Yokosawa <akiysw@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>