advsync: Replace table with diagrams in "Variables Can
 Have More ..."

In commit 8de1474e3282 ("Add larger multiple-value-write scenario"),
a table representing difference in opinions of 15 CPUs of a
shared variable was introduced. However, a diagram similar to
Figure 14.2 is nicer for readers.
This commit replaces the table with a figure which consists of
two diagrams. One shows the overall picture and the other shows
the zoom-up of the beginning period.
It also updates related text appropriately.
It removes the explanation of non-response period added in
commit 12eb7c30e5db ("Explain why cells in table 14.2 can be blank")
since the diagrams ignore such periods just as in Figure 14.2.
It also adds unpolished source code of a conversion program
used to generate .fig files.
It also adds the input file to the conversion program for
other people to see if the extracted data is correct.
Detailed explanation of the format of the input file and the
conversion program is planned to be added in a future commit.

Signed-off-by: Akira Yokosawa <akiysw@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>