advsync: Use colors nicer to your eyes
In the diagrams added in commit f324c22db701 ("advsync: Replace
table with diagrams in \"Variables Can"), fully saturated colors
were used. They are not friendly to your eyes.
This commit uses less saturated colors of blue and red hues.
It also adds trial versions of the diagrams which depict the
"non-responsive" state of CPUs, whose info were included in
the original table, by narrow bars of higher saturation colors
on the lower side.
They are not referred to in the text yet.
Input data and the source of the conversion program are also
updated.

Suggested-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>
Signed-off-by: Akira Yokosawa <akiysw@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>