.gitignore: Add planned empty targets
This change will prevent "git status" from complaining when you
switch from a planned test branch to other branches.
Note: Target names 'extraction' and 'embedfonts' are now denoted
as phony targets on the master branch. Therefore, it is safe to
change them to empty targets of the same names on the planned test
branch.

Signed-off-by: Akira Yokosawa <akiysw@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>