Add font installation check
This commit adds tests in eps2pdf.sh and runlatex.sh which check
the presence of the Steel City Comic font in the font cache and
in a PDF file generated from a .svg file.
If either of the tests fails, a message referring the user to
item 1 in FAQ.txt will be displayed.
This commit also adds a phony target 'touchsvg' in Makefile.
It is useful when a user fails to install the font before doing
"make" for the first time. It used to require a "make neatfreak"
for the installed font to actually make any effect.
This commit also adds item 1.1 in FAQ.txt which explains the use
of "make touchsvg".

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>