sig-theft: Fix .dot source for dot - graphviz version 2.36.0
The 'dot' command of Ubuntu 14.04 "graphviz version 2.36.0
(20140111.2315)" has trouble with count/sig-theft.dot.
It exits with a message:
    Warning: count/sig-theft.dot: syntax error in line 13 near ']'

This commit fixes the .dot source so that count/sig-theft.eps can
be generated as expected.

The original count/sig-theft.eps seems to have been built with
'dot' of Graphviz version 2.12. I'm not sure if that version of
'dot' can process the modified sig-theft.dot correctly.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>