Make default target of "make" overridable
This commit adds a conditional assignment in Makefile so that
the default target can be overridden by an environment variable.

If you have the environment variable "PERFBOOK_DEFAULT" set, its
value will be regarded as the default target. Otherwise,
perfbook.pdf is the default as before.

This commit also updates FAQ.txt to add an explanation of the
environment variable.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>