Further improvement of build scripts
This commit restructures Makefile so that dependencies can be
checked more rigorously and redundant command invocations
can be mostly avoided.
The following changes are done in Makefile:
    1) Use empty targets:
        perfbook_aux, extraction, embedfonts, etc.
    2) Add static pattern rules of (.tex -> .eps) and
       (.dot -> .pdf), and remove rules for individual files
    3) rm's of .eps files whose sources are .tex or .dot files
       are moved to distclean.
    4) Minor modification of variable definitions.

The change in runlatex.sh is necessary to avoid redundant
pdflatex run.
This commit also removes qqz_first in .gitignore, which was
planned to be used as an empty target but is not used right
now.

In v1 of this series, removal of qqz.tex, contrib.tex and
origpub.tex from the repository was attempted. However,
their dependency loop around perfbook_flat.tex requires them
to be kept in the repository. In addition, they are useful in
detecting possible regression caused by this commit.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>