Remove .eps files whose sources are .dot
In commit 72b41446f032 ("rm qqz.tex and add store15tred.esp to
repository"), advsync/store15tred.eps is added to the repository.
However, since we already have an implicit rule for .dot -> .eps
conversion, it is safe to remove .eps files whose sources are
.dot files. This commit removes them.

It also updates .gitignore accordingly.

As for .eps files whose sources are .tex files, they used to be
removed by 'make clean' before the cleanup of build scripts.
This commit adds them as well as the removed .eps files to the
'distclean' rule.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>