Makefile: Promote eps2pdf.sh to implicit rules
In the first build or a build after removing '.pdf' files by
'make neatfreak' or 'make distclean', '.pdf' figures can be built
in parallel if the rules for '.eps' -> 'pdf' and '.svn' -> 'pdf'
are given in Makefile instead of shell scripts.

This commit adds what are done in utilities/eps2pdf.sh, etc. as
implicit rules in Makefile.
As the result of the change, empty target 'embedfonts' is removed.
Now 'make -j2' works fine despite several warnings of duplicated
targets. Those warnings will be taken care of by a later patch.

Note that as of this commit, '-jN' option is not guaranteed to always
work as expected.
It is safe to use it when you specify a single target.

Using '-jN' for multiple targets such as in 'make -j2 1c 2c', some
of the targets may fail to build depending on the timing of parallel
processing, supposedly due to interference of '.aux' files used by
'bibtex'.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>