skiplist: Avoid self-deadlock in skiplist_insert_lock()
There is a self-deadlock in skiplist_insert_lock(), here is an example
for how the deadlock happens:

Given a skiplist with only one item, whose key is 3 and ->sl_toplevel is
2:

	.....

	+----+          +----+          +----+
2	|head|--------->|  3 |--------->|NULL|
	+----+          +----+          +----+
1	|head|--------->|  3 |--------->|NULL|
	+----+          +----+          +----+
0	|head|--------->|  3 |--------->|NULL|
	+----+          +----+          +----+

then a new item is inserted with key=5, and given 1 as ->sl_toplevel by
random_level(). skiplist_insert_lock() will get into a state as:

	level = 1

	.....						.....

	+----+          +----+          +----+          +----+
2	|head|--------->|  3 |--------->|NULL|          |NULL|
	+----+          +----+          +----+          +----+
1	|head|--------->|  3 |--------->|NULL|          |slp |
	+----+          +----+          +----+          +----+
0	|head|--------->|  3 |--------->|NULL|          |  ? |
	+----+          +----+          +----+          +----+
			   ^                            update
                           |
			  slp

where
	update[level] != update[level+1]

so
	the lock of slp will be held

then because

	level < slp->toplevel
and
	slp->slp_next[level] == slp->slp_next[level+1]

we will drop the lock and retry, which could result in the same
situation.

To fix this, we need an additional check in the retry check:

	level < toplevel

because toplevel is the highest level for the to-be-inserted item,
therefore we won't take the locks of its possible predecessors at levels
higher than toplevel, as a result, we don't need to and should not check
those levels for retry.

Signed-off-by: Boqun Feng <boqun.feng@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>