toolsoftrade: Adjust position of Figure 4.5
After commit f104c4f9ef15 ("toolsoftrade: Centering figures of
verbatim code fragments"), Figure 4.5 is positioned just below
Figure 4.4 in perfbook.pdf.
To restore the position of Figure 4.5, this commit moves its
'picture' environment slightly backward in the latex source.
It is easier to compare Figures 4.4 and 4.5 in this layout.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>