qqz: Add macro's for referencing quick quizzes
So far, reference to qqz in the text is done in a manual way.
This commit adds macro's for defining and referencing a qqz name.
The reference will also generate a hyperlink.

This commit also modifies where the referencings of qqz's are
done in count/count.tex to use the macro's.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>