qqz: Add cross-link reference at the end of each QQ and QQA
This commit adds cross-link reference to the special character
at the end of each QuickQuiz and QuickQuiz Answer.
This is convenient when you read a long QuickQuiz or its answer
and want to go back and forth between main text and Appendix D.
For QuickQuiz Answer, a special character distinct from the one
at the end of QuickQuiz is added at the end for this purpose.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>