qqz: Improve accuracy of qqz back link
This commit improves the accuracy of qqz back link by moving
\label{} command just after "Quick Quiz x.y:".

Before this commit, some Quick Quizzes are placed at the top
of columns (or pages), and the back links to them from their
answers point the bottom of the previous columns (or pages).
After this commit, Quick Quizzes seem not to be placed at the
top, and the back links always point the right points.

This means that column and page breaks are done at slightly
different positions in the resulting pdf.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>