epigraph: Use \epigraphhead{}
This commit makes use of \epigraphhead{} so that epigraphs are
placed in header areas of chapter heading pages.
To make the change less invasive, it defines a macro \Epigraph{}
in the preamble and use it after each chapter heading.
By modifying the definition of the macro, it would be easy to
tweak layout of epigraphs.

Note that no text (nor line break) is permitted between \chapter{}
and \epigraphhead{} commands. So each blank line above \epigraph{}
is replaced with a comment.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>