Disable 'microtype' in code snippets
"microtype" package can affect monospece-ness of code snippets.
The symptom is evident in Figures 12.17 and 12.18, where patch
diffs beginning with "-" and "+" have slightly different indents.
This commit addresses this issue by adding a \microtypesetup{}
command at the start of "verbatim" and "verbbox" environments
with the help of "etoolbox" package.

This workaround is mentioned in Section 9 of microtype documentation:
http://mirrors.ctan.org/macros/latex/contrib/microtype/microtype.pdf

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>