Makefile: Add target '1csf' (one-column sans serif)
Sans serif fonts are said to be easier to read on a (not high
definition) display than serif fonts.
Current perfbook-1c.pdf is supposed to be for such readers, but
is using serif font (Times Roman).
To try a sans serif font (Helvetica) version for one-column layout,
this commit adds a new boolean "sansserif" in the preamble and
a make target "1csf" that sets the boolean true.

Since the boolean is referred to in an earlier part of the preamble,
declarations of custom booleans are moved forward along with the
new boolean.

The result can not be said as optimal, but it is mostly good.
As for monospece font this commit uses "inconsolata", which is
mostly sans serif.
For epigraphs, Times Roman is used as before.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>