Abort build if 'mpost' is present in sources
A POC of an arbitrary command execution vulnerability in TeX config
was disclosed on November 28, 2016.

To avoid the exploitation, this commit adds checks just before latex
invocations in runfirstlatex.sh, runlatex.sh, and the ".tex -> .eps"
rule of Makefile.

The added script "mpostcheck.sh" first checks if your config is
secure. If it is not, then it aborts the build if a word "mpost" is
found in the sources. (It is not present at the moment.)

If "mpost" is not found, the build will continue. But mpostcheck.sh
displays a warning message. To make sure the message can catch eyes,
the script is also invoked at the end of runlatex.sh.

Once you fix your TeX config, the behavior of the build scripts
will be the same as before.

Refer to the comment in mpostcheck.sh for the circumstances and
instructions for fixing your TeX config.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>