Increase floatpagefraction for 2c layout
In 2c layout, it sometimes looks imbalanced when a floating
figure or table with the "[p]" option occupies a column while
its next column is flushed from top to bottom. This is because
"floatpagefraction" command is defined as .5 (a figure or a table
whose height is 50% or more of textheight can occupy a column).

To prevent such occupations from happening in 2c layout, this
commit sets floatpagefraction as the same as "topfraction" command
for 2c.

This change will reduce typeset freedom of TeX engine, and might
cause some floating objects to be placed further from where they
are referenced in the text. If this causes an undesirable result,
we might need to reduce the value of floatpagefraction.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>