Define \tco{} command to be used in table
The \co{} command interferes with column width estimation of
tabular environment. To avoid the issue, this commit defines
a command "\tco{}" to be used instead. It will still break at white
spaces, but not at "_".
This commit also updates the definition of \nbco{} to make sure
that it won't break at "_".

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>