alphapf.bst: Make use of 'location' field in 'inproceedings' entry
In early "inproceedings" entries, "address" fields are used to
indicate the location of conferences as is described in BibTeX
documentation[1].

However, in recent entries such as those copied from e.g. ACM
BibTeX templates, "address" fields are used for publisher's
addresses. In such entries, "location" fields are used instead
to show where those events were held.

The circumstances of how to handle locations of conferences in
bibliography databases are far from being settled within LaTeX
community[2][3].

This commit therefore adds awareness of "location" field in
alphapf.bst for "inproceedings" entry, and lets the field have
a higher precedence than an "address" field if both fields are
present. It has no effect on other types of entries or
"inproceedings" entries with "address" field only.

This change can be thought of as a compromise to improve the
consistency presented in Bibliography in a non-intrusive way.

This commit also adds comment in the header describing changes made
for perfbook.

  [1] Excerpt of BibTeX documentation (shown by "texinfo bibtex"):

    10. The PROCEEDINGS and INPROCEEDINGS entry types now use the
        address field to tell where a conference was held, rather
        than to give the address of the publisher or organization.
        If you want to include the publisher's or organization's
        address, put it in the publisher or organization field.

  [2] "How to show the location of a conference in BibTeX":
      http://tex.stackexchange.com/questions/76566/

  [3] "Abuse of address field in BibTeX entries":
      http://tex.stackexchange.com/questions/152725/

Note: It is easier to see the diff with white-space changes ignored.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>