appendix: Draw hyperlink border in bibliography for 1c layout
Reflecting recent addition of URLs and DOIs in bibliography,
"pdfborder" would be helpful to find embedded links.

As an experiment, this commit enables it for 1c layout.
The color of the border is tuned to "light blue" to be nice for
reader's eyes.

Note: This commit involves indent changes. diff is easier to see
by ignoring white space changes.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>