appendix: Enable hyperlink border in bibliography for all layouts
"pdfborder" in bibliography also helps in identifying known broken
urls. This commit enables the border regardless of layout.

If a hyperlink crosses a page boundary, the hyperlink is wrongly embedded
also in the footer and header strings. As of now, target "1csf" hits this
issue at the bottom of first page of Bibliography (p. 653-654).
This is a well known issue of "hyperref" package, which would cause
a build error in 2c layout. Before enabling the border, we just didn't
notice the issue.

Suggested-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>
Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>