Choose 'cmtt' font for code snippets in targets 'msnt' and 'mstx'
By specifying "cmtt" (Computer Modern Teletype) font as \ttdefault
in verbatim, verbbox, table, tabular, tabulary, and minipage
environments, we can improve consistency of looks independent of
monospace font choice for main text. cmtt font is more than suitable
for listing of code snippet and program output.

However, when mixed used with Times Roman (clone) font, cmtt's
slightly lower x-height than Times Roman, and its similar thickness
as Times Roman, it is not distinctive enough as can be seen in
"mslm" target.

In this respect, "txtt" and its enhanced version "newtxtt" fonts
are more suitable as is described in [1].

This commit enables the font choice of cmtt for code snippets in
targets "msnt" and "mstx".

Help messages in Makefile are updated accordingly.

Any feedback on the font choice is welcome.

[1] Excerpt from newtxtt documentation:

    This package provides an interface to another alternative---the
    typewriter fonts provided with txfonts, with some enhancements.
    They have the same widths as cmtt, but are taller, heavier,
    more geometric and less shapely, with very low contrast, and
    are more suited to match Roman fonts of height and weight
    approximating that of Times.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>