advsync: Adjust spacing in Tables 14.1 and 14.2, consistent indirection
Make the mentions of litmus-test variables omit the indirection, which
matches the "exists" clause and shortens things a bit.  Note that
the litmus-test function parameter lists include the indirection and
the exists clause omits it, so this is a free choice: Equal confusion
either way.  So come down on the side of brevity.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
[ paulmck: Match exists clause rather than function arguments. ]
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>