Use 'small' font in caption of code snippet
As code snippets use "scriptsize" font, the "normal" font of their
captions looks too big when placed just above the code in the "ruled"
style. Using the "small" font can reduce the size gap.

Also use colon as a separator in the caption as the same as those of
figures and tables.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>