styleguide: Reflect recent conversion of code snippets to listing
Also use a different approach to reduce skips of table captions.
By setting a smaller value to \abovecaptionskip, captions of tables
and figures can be made closer to the body regardless of their
positions and avoid ambiguity of them to look like captions
to other nearby tables/figures.

In the experimental tables, a negative value is still set
to \abovetopsep to further reduce the skips.

As options of \captionsetup{} related to vertical spacing are
ignored in captions controlled by the "float" package, remove them.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>