locking: Fix Conditional Locking to Reduce Contention
The "Conditional Locking to Reduce Contention" listing contains
a needless check of gp_flags after the root ->fqslock has been
acquired.  However, in real life, this check is helpful.  This
commit therefore changes the example to require invoking
do_force_quiescent_state() at most once per 100 milliseconds,
which means that the winner of the tournament needs to recheck
in order to see if the previous winner won within the pase
100 milliseconds.

Reported-by: Yubin Ruan <ablacktshirt@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>