Tweak figure size
Some of the figures are rendered too large or too small.
This commit tweaks argmuents of \resizebox{} macros to make those
figures look consistent in their sizes.

Two figures in advsync/rt.tex was too small with the column width
of 2c layout. They are converted to "figure*" environments with
larger widths.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>