Makefile: Support URW++ v2 font family names
Add conversion scripts for URW++ v2 font family, which has been
introduced in Fedora 27.

Define macros in Makefile according to the output of "fc-list" so
that correct scripts are used for font conversion.

NOTE: Ghostscript 9.22 is not compatible with a2ping at the moment.[1]
Downgrading to 9.20 is necessary on Fedora 27.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1530268

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>