future/QC: Improve layout of Table 17.5 by using subfloat
This enables us to offload part of row headers to captions of
sub-tables.

In perfbook.tex, font size of sub captions is increased to "small"
from the default of "footnotesize".

Also remove "%" in each cell and indicate the unit in row headers.
By this change, the "S" column-layout specifier of siunitx package
can be used to align decimal points.

Also add headers to indicate what "Qn" and "CXm_n" mean.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>