future/QC: Tweak horizontal spacing of Table 17.5
By employing \multirow command of the "multirow" package, the width
of the 1st column of entanglement table can be reduced.

Note: The 1st argument of \multirow specifies the number of rows to span,
and the 2nd argument, in this case {1in}, specifies the width of the
column. Internally it is translated to a \parbox.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>