Add poor person's slanted shape for 'newtxtext' font
The "newtx" font package switched to TeX Gyre Termes fonts in
version 1.4 (released in TeX Live 2014). It lacks slanted shape.

In perfbook, slanted shape is not used in LaTeX sources at the
moment, but is used in page headers generated in the "book" document
class.

The change of Times Roman clone font to "newtxtext" font made in
commit b4ad25eae241 ("future/QC: Use upright glyph for math constant
and descriptive suffix") had a side-effect of page headers to
be rendered in italic font when TeX Live 2014 or later is used.
Most notably, chapter/section numbers in page headers are rendered
upright.

This commit adds poor person's slanted shapes for roman fonts
as suggested in https://tex.stackexchange.com/questions/247539/.

This change doesn't have visual effect on TeX Live 2013.

Fixes: b4ad25eae241 ("future/QC: Use upright glyph for math constant and descriptive suffix")
Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>