Stop using modified book.cls
Strictly speaking, using a modified version of a standard class
file should be avoided. Instead, define a class "pfbook" in
pfbook.cls which loads the book class to add an option and
override a couple of definitions.

The addition of the option was originally done in commit
7b8ce4d5ece9 ("Modify book.cls to allow 10.75x8.25-inch PDFs
for hardcover use").

The modification of definitions was originally done in commit
398070c4ae9a ("book.cls: Fix indent after \chapter in twocolumn").

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>