toolsoftrade: Add labels in code samples as comments
For referencing from text, put labels at the end of lines as
comments in code samples.

Labels look like the following (in the case of C source):

  ... //%lnlbl[<line label>]

, or

  ... //$lnlbl@<line label>^

, or other choices of escape charactors depending to
the charactor set used in a particular code snippet.

In the first example, "%", "[", and "]" are the escape
charactors to be replaced with "\", "{", and "}", respectively.
In the second example, "$", "@", and "^" are selected as
escape charactors.

By escape-to-LaTeX feature of "fancyvrb"
package, both examples will be equivalent to:

  ... \lnlbl{<line label>}

with " //" removed by a conversion script (to be added
in a following commit) in "Verbatim" environment of LaTeX
sources.

\lnlbl{} macro is to be defined in preamble of perfbook.tex
so that the actual label has a form with a prefix string
"ln:<chapter>:<file name>:<function>:".

Any sets of three charactors are good as escape charactors
as long as they don't appear in a code snippet.
Other candidates of escape charactors would be:

    "\", "#", "!", "|", "`", "~"

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>