toolsoftrade: Reference line in code snippets by label
The "fancyvrb" package provides verbatim like environment with
escape-to-LaTeX feature.  This commit attempts to apply this
scheme and to update build scripts.

Suffix of ".fcv" is chosen for snippets extracted by fcvextract.pl.
Extracted snippets are placed in the same directory under
CodeSamples/, and have names "<filename>@<function>.fcv".
It would be better if we can use "<filename>:<function>.fcv",
but ":" in file names are hard to handle in Makefile targets.

CodeSamples/snippets.mk is automatically generated using
utilities/gen_snippets_mk.pl. when necessary.

In perfbook.tex, "VerbatimL" environment is defined by using
\DefineVerbatimEnvironment{} command to enable line numbering.
Automatic tab expansion (tab -> 2 spaces) and narrow left margin
are also enabled.

We define \lnlbl{} and \lnref{} macros to be used for labeling
and referencing of lines. To suppress hyperlinks from line numbers
into code snippets, \ref*{} command is used for \lnref{}.
Prefixes defined as macros \lnlblbase and \lnrefbase are concatinated
to line labels.

The Verbatim environment is supposed to be used in running text
and it adds extra vertical space at the end. The space is redundant
for VerbatimL which is used within listing environment.
To compensate, a negative "\vspace*{}" command is appended using
"\AfterEndEnvironment{}" command.

"VerbatimM" environment is also defined in the preamble for use in
inline code snippets. It is similar to "VerbatimL" environment but
with a wider left mergin and vertical space around it.

Note that "\centering" has no effect for VerbatimL environment.
Now that captions to Listings are left-aligned, centering code
snippets has no merit any more.

Several inline and floating code snippets in toolsoftrade have
been converted to the new scheme to demonstate how it work.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>